Move routing forms outside of ee and use LicenseRequired component instead. (#7639)
