+++ packages/app-store/apps.server.generated.ts
@@ -10,7 +10,6 @@ export const apiHandlers = {
   campfire: import("./campfire/api"),
   closecom: import("./closecom/api"),
   cron: import("./cron/api"),
-  "routing-forms": import("./ee/routing-forms/api"),
   exchange2013calendar: import("./exchange2013calendar/api"),
   exchange2016calendar: import("./exchange2016calendar/api"),
   exchangecalendar: import("./exchangecalendar/api"),
@@ -34,6 +33,7 @@ export const apiHandlers = {
   rainbow: import("./rainbow/api"),
   raycast: import("./raycast/api"),
   riverside: import("./riverside/api"),
+  "routing-forms": import("./routing-forms/api"),
   salesforce: import("./salesforce/api"),
   sendgrid: import("./sendgrid/api"),
   signal: import("./signal/api"),
