+++ packages/app-store/routing-forms/DESCRIPTION.md
@@ -1,8 +1,8 @@
 ---
 items:
-  - /api/app-store/ee/routing-forms/1.jpg
-  - /api/app-store/ee/routing-forms/2.jpg
-  - /api/app-store/ee/routing-forms/3.jpg
+  - /api/app-store/routing-forms/1.jpg
+  - /api/app-store/routing-forms/2.jpg
+  - /api/app-store/routing-forms/3.jpg
 ---
 
 It would allow a booker to connect with the right person or choose the right event, faster. It would work by taking inputs from the booker and using that data to route to the correct booker/event as configured by Cal user
