+++ packages/app-store/routing-forms/config.json
@@ -4,13 +4,14 @@
   "title": "Routing Forms",
   "slug": "routing-forms",
   "type": "routing-forms_other",
-  "imageSrc": "/api/app-store/ee/routing-forms/icon.svg",
-  "logo": "/api/app-store/ee/routing-forms/icon.svg",
+  "imageSrc": "/api/app-store/routing-forms/icon.svg",
+  "logo": "/api/app-store/routing-forms/icon.svg",
   "url": "https://cal.com/apps/routing-forms",
   "variant": "other",
   "categories": ["other"],
   "publisher": "Cal.com",
   "email": "help@cal.com",
+  "licenseRequired": true,
   "description": "It would allow a booker to connect with the right person or choose the right event, faster. It would work by taking inputs from the booker and using that data to route to the correct booker/event as configured by Cal user",
   "__createdUsingCli": true
 }
