+++ packages/app-store/routing-forms/lib/getFieldIdentifier.ts
