+++ packages/emails/src/templates/index.ts
@@ -19,4 +19,4 @@ export { OrganizerRescheduledEmail } from "./OrganizerRescheduledEmail";
 export { OrganizerScheduledEmail } from "./OrganizerScheduledEmail";
 export { TeamInviteEmail } from "./TeamInviteEmail";
 export { BrokenIntegrationEmail } from "./BrokenIntegrationEmail";
-export * from "@calcom/app-store/ee/routing-forms/emails/components";
+export * from "@calcom/app-store/routing-forms/emails/components";
